entities move better

+++ controller/Move.hpp
@@ -17,8 +17,8 @@ public:
     explicit Move(const std::shared_ptr<Humanoid>& humanoid, unsigned desiredX, unsigned desiredY);
     void execute(Field& field) override;
 private:
-    int speedX;
-    int speedY;
+    long speedX;
+    long speedY;
 };
